requirements.txt
fonttools
numpy
opencv-python
pillow >= 8.0.0
PySimpleGUI